+++ src/main/java/org/folio/entitlement/service/EntitlementModuleService.java
@@ -42,6 +42,16 @@ public class EntitlementModuleService {
       .toList();
   }
 
+  @Transactional(readOnly = true)
+  public boolean isEntitlementExist(String moduleId) {
+    return repository.existsByModuleId(moduleId);
+  }
+
+  @Transactional(readOnly = true)
+  public boolean isNoOtherEntitlementExist(String moduleId, UUID tenantId) {
+    return !repository.existsByModuleIdAndTenantIdNot(moduleId, tenantId);
+  }
+
   public void save(ModuleRequest moduleRequest) {
     var entity = mapper.map(moduleRequest);
     repository.save(entity);