+++ src/test/java/org/folio/entitlement/it/FlowExecutionImmediateTimeoutIT.java
@@ -40,7 +40,7 @@ import org.springframework.test.web.servlet.ResultActions;
 @SqlMergeMode(MERGE)
 @Sql(executionPhase = AFTER_TEST_METHOD, scripts = "classpath:/sql/truncate-tables.sql")
 @TestPropertySource(properties = {
-  "application.kong.enabled=false",
+  "application.apigw.enabled=false",
   "application.keycloak.enabled=false",
   "application.flow-engine.execution-timeout=1ms",
   "application.clients.folio.connect-timeout=60s",